requirements.txt
pyTelegramBotAPI
google-generativeai
requests
beautifulsoup4